disable alarms, change temp conv. rate, new alarm
-Added function to disable alarms. This can be used when stopping the
logger with an external interrupt triggered by a button.

- Added function to set temperature conversation rate in RTC. Default is
64 seconds. This function will update it to 512 seconds (~8.53 mins) to
increase battery life of the coin cell. Might reduce timekeeping
accuracy in environments with rapid temperature fluctuations.

- Added function to set new alarm. This can be called after alarms were
first disabled. This function takes the current time of the RTC and sets
the alarm after a specified interval, which is passed as an argument.